Update cuda-ubuntu.dockerfile
on devieQuery build step, you must run 'cmake .' prior to  'make' command, since there is no makefile until you do that. 
The 'cmake' binary is not part of the initial install so that must be added to apt install in previous step.
This step will fail without this, and halt the image building process.